#6 [refactor] : package 정리
