+++ app/src/main/java/org/sopt/and/feature/signup/SignUpViewModel.kt
@@ -1,11 +1,11 @@
-package org.sopt.and.viewmodel
+package org.sopt.and.feature.signup
 
 import android.util.Patterns
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
-import org.sopt.and.presentation.utils.AuthKey.PASSWORD_PATTERN
+import org.sopt.and.utils.AuthKey.PASSWORD_PATTERN
 
 class SignUpViewModel : ViewModel() {
     var email by mutableStateOf("")
